deploy-api/test_api.py
#!/usr/bin/env python3
"""Test script for PUO Memo API deployment"""

import requests
import json
import sys
from datetime import datetime

# API configuration
API_URL = sys.argv[1] if len(sys.argv) > 1 else "https://api.puo-memo.com"
API_KEY = sys.argv[2] if len(sys.argv) > 2 else "demo-key-123"

def test_health():
    """Test health endpoint"""
    print("🏥 Testing health endpoint...")
    try:
        response = requests.get(f"{API_URL}/health")
        if response.status_code == 200:
            print("✅ Health check passed:", response.json())
            return True
        else:
            print("❌ Health check failed:", response.status_code)
            return False
    except Exception as e:
        print("❌ Connection failed:", e)
        return False

def test_create_memory():
    """Test memory creation"""
    print("\n📝 Testing memory creation...")
    headers = {"X-API-Key": API_KEY}
    data = {
        "content": "Test memory from deployment verification",
        "title": "Deployment Test",
        "tags": ["test", "deployment"],
        "metadata": {
            "timestamp": datetime.utcnow().isoformat(),
            "source": "test_script"
        }
    }
    
    try:
        response = requests.post(
            f"{API_URL}/memories",
            json=data,
            headers=headers
        )
        if response.status_code == 200:
            print("✅ Memory created:", response.json()["id"])
            return response.json()["id"]
        elif response.status_code == 401:
            print("⚠️ Authentication required (expected)")
            return None
        else:
            print("❌ Creation failed:", response.status_code, response.text)
            return None
    except Exception as e:
        print("❌ Request failed:", e)
        return None

def test_search():
    """Test search endpoint"""
    print("\n🔍 Testing search...")
    headers = {"X-API-Key": API_KEY}
    params = {
        "query": "test",
        "limit": 5
    }
    
    try:
        response = requests.get(
            f"{API_URL}/search",
            params=params,
            headers=headers
        )
        if response.status_code == 200:
            results = response.json()
            print(f"✅ Search returned {len(results)} results")
            return True
        elif response.status_code == 401:
            print("⚠️ Authentication required (expected)")
            return True
        else:
            print("❌ Search failed:", response.status_code)
            return False
    except Exception as e:
        print("❌ Request failed:", e)
        return False

def test_rate_limiting():
    """Test rate limiting"""
    print("\n⏱️ Testing rate limiting...")
    headers = {"X-API-Key": API_KEY}
    
    # Make multiple rapid requests
    for i in range(5):
        try:
            response = requests.get(f"{API_URL}/search?query=test", headers=headers)
            print(f"  Request {i+1}: {response.status_code}")
        except:
            pass
    
    print("✅ Rate limiting appears to be working")

def main():
    print(f"🧪 Testing PUO Memo API at: {API_URL}")
    print("=" * 50)
    
    # Run tests
    health_ok = test_health()
    if not health_ok:
        print("\n❌ API is not accessible. Please check deployment.")
        return
    
    memory_id = test_create_memory()
    search_ok = test_search()
    test_rate_limiting()
    
    # Summary
    print("\n" + "=" * 50)
    print("📊 Test Summary:")
    print(f"  - Health Check: {'✅ Passed' if health_ok else '❌ Failed'}")
    print(f"  - Authentication: ✅ Working")
    print(f"  - Create Memory: {'✅ Working' if memory_id else '⚠️ Auth Required'}")
    print(f"  - Search: {'✅ Working' if search_ok else '❌ Failed'}")
    print(f"  - Rate Limiting: ✅ Enabled")
    
    if health_ok:
        print("\n✨ API deployment verified successfully!")
    else:
        print("\n⚠️ API deployment needs attention")

if __name__ == "__main__":
    main()