deploy-api/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
python-dotenv==1.0.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# API utilities
httpx==0.25.2
slowapi==0.1.9
python-multipart==0.0.6

# AI/ML
google-generativeai==0.3.2
numpy==1.26.4
scikit-learn==1.4.0
tiktoken==0.5.2

# Monitoring
prometheus-client==0.19.0

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Authentication
supabase==2.0.0
email-validator==2.1.0